+++ backend/types/database/topic.go
@@ -10,7 +10,7 @@ type Topic struct {
 
 type Card struct {
 	Id       uint64        `json:"id"`
-	Score    float64       `json:"score"`
+	Score    int32         `json:"score"`
 	Opened   bool          `json:"opened"`
 	Question string        `json:"question"`
 	Duration time.Duration `json:"duration"`
